restructuring: introduce bexpr
Still to do
* Move relabel and subst from Behave to BExpr
* use weeder
