+++ stack_linux.yaml
@@ -43,6 +43,7 @@ packages:
 - location: sys/behave
 - location: sys/behavedefs
 - location: sys/behaveenv
+- location: sys/bexpr
 - location: sys/cnect
 - location: sys/core
 - location: sys/coreenv
