+++ sys/behave/behave.cabal
@@ -18,7 +18,6 @@ library
   exposed-modules:      Behave
                       , Expand
                       , Subst
-                      , StautDef
 
   other-modules:        Equiv
                       , Next
@@ -35,6 +34,7 @@ library
 
                       , behavedefs
                       , behaveenv
+                      , bexpr
                       , coreenv
                       , defs
                       , solve
@@ -73,6 +73,7 @@ test-suite behave-test
                      
                      , behavedefs
                      , behaveenv
+                     , bexpr
                      , coreenv
                      , defs
                      , front
