+++ sys/bexpr/Setup.hs
@@ -1,0 +1,7 @@
+{-
+TorXakis - Model Based Testing
+Copyright (c) 2015-2017 TNO and Radboud University
+See LICENSE at root directory of this repository.
+-}
+import Distribution.Simple
+main = defaultMain
