+++ sys/bexpr/bexpr.cabal
@@ -1,0 +1,54 @@
+name:                   bexpr
+version:                0.0.1.0
+synopsis:               Behaviour Expressions
+license:                BSD3
+license-file:           license.txt
+author:                 Jan Tretmans
+                      , Pierre van de Laar
+maintainer:
+                      , pierre.vandelaar@tno.nl
+copyright:              TNO
+category:               Web
+build-type:             Simple
+cabal-version:          >=1.10
+
+library
+  hs-source-dirs:       src
+
+  exposed-modules:      BehExprDefs
+                      , ChanId
+                      , ProcDef
+                      , ProcId
+                      , StautDef
+                      , StatId
+                      , VarEnv
+
+  ghc-options:         -Wall -Werror -O2 -optc-O3 -optc-ffast-math
+
+  build-depends:        base
+                      , containers
+                      , deepseq
+                      , text
+                      
+                      , valexpr
+
+  default-language:     Haskell2010
+  
+test-suite bexpr-test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+                     
+  main-is:             TestMain.hs
+  
+  other-modules:       TestBehExprDefs
+  
+  ghc-options:         -Werror -Wall -O2 -optc-O3 -optc-ffast-math -threaded -rtsopts -with-rtsopts=-N                     
+  
+  build-depends:       base
+                     , containers
+                     , QuickCheck
+
+                     , bexpr
+                     , valexpr
+
+  default-language:    Haskell2010
